finishing code structure
