+++ rnote-ui/src/workspacebrowser/filerow/widget_helper/mod.rs
@@ -1,0 +1,1 @@
+pub mod entry_dialog;
